Remove `bytes` feature, make `concat_bytes!` always available
This feature was only really required for controlling how the tests were
compiled. It shouldn't be public API. The macro is always available but
the `feature(concat_bytes)` is required to use it without compiler
errors. This change makes the `bytes` feature private API (now `_bytes`)
and only uses it in tests.